drop out net

+++ Page5/DogVsCat/ProcessData.py
@@ -10,6 +10,8 @@ class PreProcessData:
         print(INFO + 'Processing Start.')
         # original_data_dir = r'/Users/Yuseng/Downloads/Deep-Learning-For-Computer-Vision-master/datasets/animals'
         original_data_dir = r'/Users/Yuseng/Downloads/all/train'
+        # original_data_dir = r'/home/bigdata/Documents/DeepLearningProject/CatVsDog/train'
+        # original_data_dir = r'/Users/zzc20160628-14/Downloads/cat_dog_data/train'
         base_dir = './cat_and_dog_small'
         ut.ifNoneCreateDirs(base_dir)
 
@@ -46,6 +48,8 @@ class PreProcessData:
             if not os.path.exists(src):
                 raise FileNotFoundError
             dst = os.path.join(cat_train_dir, name)
+            if os.path.exists(dst):
+                continue
             shutil.copy(src=src, dst=dst)
 
         fnames = ['cat.{}.jpg'.format(i) for i in range(1000, 1500)]
@@ -54,6 +58,8 @@ class PreProcessData:
             if not os.path.exists(src):
                 raise FileNotFoundError
             dst = os.path.join(cat_val_dir, name)
+            if os.path.exists(dst):
+                continue
             shutil.copy(src=src, dst=dst)
 
         fnames = ['cat.{}.jpg'.format(i) for i in range(1500, 2000)]
@@ -62,6 +68,8 @@ class PreProcessData:
             if not os.path.exists(src):
                 raise FileNotFoundError
             dst = os.path.join(cat_test_dir, name)
+            if os.path.exists(dst):
+                continue
             shutil.copy(src=src, dst=dst)
 
         fnames = ['dog.{}.jpg'.format(i) for i in range(1000)]
@@ -70,6 +78,8 @@ class PreProcessData:
             if not os.path.exists(src):
                 raise FileNotFoundError
             dst = os.path.join(dog_train_dir, name)
+            if os.path.exists(dst):
+                continue
             shutil.copy(src=src, dst=dst)
 
         fnames = ['dog.{}.jpg'.format(i) for i in range(1000, 1500)]
@@ -78,6 +88,8 @@ class PreProcessData:
             if not os.path.exists(src):
                 raise FileNotFoundError
             dst = os.path.join(dog_val_dir, name)
+            if os.path.exists(dst):
+                continue
             shutil.copy(src=src, dst=dst)
 
         fnames = ['dog.{}.jpg'.format(i) for i in range(1500, 2000)]
@@ -86,6 +98,8 @@ class PreProcessData:
             if not os.path.exists(src):
                 raise FileNotFoundError
             dst = os.path.join(dog_test_dir, name)
+            if os.path.exists(dst):
+                continue
             shutil.copy(src=src, dst=dst)
 
         print(INFO + 'Processing End.')
